[Refactor] Move PublishVersionManager ownership to agent
