+++ be/src/agent/CMakeLists.txt
@@ -24,6 +24,7 @@ ADD_BE_LIB(Agent
         task_worker_pool.cpp
         utils.cpp
         publish_version.cpp
+        publish_version_manager.cpp
         finish_task.cpp
         report_task.cpp
         agent_task.cpp
