+++ be/src/agent/agent_server.h
@@ -44,6 +44,7 @@
 namespace starrocks {
 
 class ExecEnv;
+class PublishVersionManager;
 class Status;
 class TAgentTaskRequest;
 class TAgentResult;
@@ -77,6 +78,8 @@ public:
     //
     // Returns nullptr if `type` is not a valid value of `TTaskType::type`.
     ThreadPool* get_thread_pool(int type) const;
+
+    PublishVersionManager* publish_version_manager() const;
 
     // Dedicated pool for per-file copy in lake-to-lake replication. Returned pool is distinct
     // from `get_thread_pool(TTaskType::REPLICATE_SNAPSHOT)` so that the outer agent task can
